Remove repo from build key
If a commit appeared in two repositories at the same time, this caused
us to attempt two builds. Which shouldn't be a problem, because BuildKit
should notice and combine them. Unfortunately, it only bothers to send
the build logs to one of them, leading to frequent "BuildKit missing
output" errors.